Move HmacSha1 from util package to crypto package
For more logical organization.
HmacSha1 is used to verify the integrity and authenticity of requests from GitHub.
